requirements.txt
Django==3.1
Pillow==7.2.0
django-fontawesome-5==1.0.18